Fix cardinalities.

+++ src/execution/exec/execution_context.cpp
@@ -74,7 +74,7 @@ struct features {
   uint32_t est_output_rows[MAX_FEATURES];
   uint16_t key_sizes[MAX_FEATURES];
   uint8_t num_keys[MAX_FEATURES];
-  uint8_t est_cardinalities[MAX_FEATURES];
+  uint32_t est_cardinalities[MAX_FEATURES];
   uint8_t mem_factor[MAX_FEATURES];
   uint8_t num_loops[MAX_FEATURES];
   uint8_t num_concurrent[MAX_FEATURES];
@@ -101,7 +101,7 @@ void ExecutionContext::EndPipelineTracker(const query_id_t query_id, const pipel
     feats.est_output_rows[i] = static_cast<uint32_t>(op_feature.GetNumRows());
     feats.key_sizes[i] = static_cast<uint16_t>(op_feature.GetKeySize());
     feats.num_keys[i] = static_cast<uint8_t>(op_feature.GetNumKeys());
-    feats.est_cardinalities[i] = static_cast<uint8_t>(op_feature.GetCardinality());
+    feats.est_cardinalities[i] = static_cast<uint32_t>(op_feature.GetCardinality());
     feats.mem_factor[i] = static_cast<uint8_t>(op_feature.GetMemFactor() * UINT8_MAX);
     feats.num_loops[i] = static_cast<uint8_t>(op_feature.GetNumLoops());
     feats.num_concurrent[i] = static_cast<uint8_t>(op_feature.GetNumConcurrent());
